refactor: domain 패키지와 view 패키지로 각각 분리하고, 의존성을 제한한다
