+++ README.md
@@ -6,6 +6,7 @@
 - Step 2 : [https://github.com/next-step/kotlin-racingcar/pull/1320](https://github.com/next-step/kotlin-racingcar/pull/1320)
 - Step 3 : [https://github.com/next-step/kotlin-racingcar/pull/1434](https://github.com/next-step/kotlin-racingcar/pull/1434)
 - Step 4 : [https://github.com/next-step/kotlin-racingcar/pull/1504](https://github.com/next-step/kotlin-racingcar/pull/1504)
+- Step 5 : [https://github.com/next-step/kotlin-racingcar/pull/1530](https://github.com/next-step/kotlin-racingcar/pull/1530)
 
 <br>
 
@@ -157,6 +158,7 @@
   - [x] 콘솔 출력 메시지와 입력 받는 행위는 입력 뷰 객체로 응집한다.
   - [x] 자동차 게임 객체의 UI 객체 의존성을 제거한다.
   - [x] 자동차 게임 객체의 단위 테스트를 추가한다. (Random 요소 추출)
-- [ ] domain 패키지와 view 패키지로 각각 분리하고, 의존성을 제한한다.
-- [ ] 테스트 불가능한 영역을 분리한다.
-- [ ] 테스트 가능한 모든 영역에 단위 테스트를 진행한다.
+  - [ ] 자동차 운전 정책에서 number를 사용하지 않아도 동작하도록 변경한다.
+- [x] domain 패키지와 view 패키지로 각각 분리하고, 의존성을 제한한다.
+- [x] 테스트 불가능한 영역을 분리한다.
+- [x] 테스트 가능한 모든 영역에 단위 테스트를 진행한다.
